// src/app/interfaces/used-material.ts
export interface IUsedMaterial {
      readonly Factory: string
      readonly PlanGroupID: string
      readonly PlanID: string
      MaterialCode: string
      MaterialName: string
      LotNo: string
      UsedQTY: number
      RecentlyStockQTYChangeDate: Date
      readonly UpdateDisplay: string
      readonly String1: string
      readonly String2: string
      readonly String3: string
      readonly String4: string
      readonly String5: string
      readonly Numeric1: number
      readonly Numeric2: number
      readonly Numeric3: number
      readonly Numeric4: number
      readonly Numeric5: number
      InsertUserID: string
      readonly InsertDate:string
      UpdateUserID: string
      readonly UpdateDate: string
      readonly UpdateCounter: number
}